feat(testing-instrument): add mock activity result
